third_party/WebKit/LayoutTests/accessibility/disabled-controls-not-focusable-expected.txt

This test makes sure that disabled form controls are not exposed as focusable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

button
PASS axElement.isFocusable is true
PASS axElement.isFocusable is false
text
PASS axElement.isFocusable is true
PASS axElement.isFocusable is false
checkbox
PASS axElement.isFocusable is true
PASS axElement.isFocusable is false
radio
PASS axElement.isFocusable is true
PASS axElement.isFocusable is false
submit
PASS axElement.isFocusable is true
PASS axElement.isFocusable is false
slider
PASS axElement.isFocusable is true
PASS axElement.isFocusable is false
combobox
PASS axElement.isFocusable is true
PASS axElement.isFocusable is false
listbox
PASS axElement.isFocusable is true
PASS axElement.isFocusable is false
textarea
PASS axElement.isFocusable is true
PASS axElement.isFocusable is false
PASS successfullyParsed is true

TEST COMPLETE
